% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{button}
\alias{button}
\title{Create a value box component for a dashboard.}
\usage{
button(value, color = NULL, href = NULL)
}
\arguments{
\item{value}{the text in the button}

\item{color}{Background color for the button. This can be one of the built-in
background colors ("primary", "info", "success", "warning", "danger") or
any valid CSS color value.}

\item{href}{An optional URL to link to. Note that this can be an anchor of
another dashboard page (e.g. "#details").}
}
\description{
A value box displays a value (usually a number) in large text, with a smaller
caption beneath, and a large icon on the right side.
}
\details{
See the flexdashboard website for additional documentation:
 \href{http://rmarkdown.rstudio.com/flexdashboard/using.html#value_boxes}{http://rmarkdown.rstudio.com/flexdashboard/using.html#value_boxes}
}
\examples{
library(flexdashboard)

button('Click here')
}
