% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_dashboard.R
\name{flex_dashboard}
\alias{flex_dashboard}
\title{R Markdown Format for Flexible Dashboards}
\usage{
flex_dashboard(fig_width = 5, fig_height = 3.5, dev = "png",
  smart = TRUE, self_contained = TRUE, fill_page = FALSE,
  orientation = c("columns", "rows"), theme = "default",
  highlight = "default", mathjax = "default", css = NULL,
  includes = NULL, lib_dir = NULL, md_extensions = NULL,
  pandoc_args = NULL, devel = FALSE, ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{smart}{Produce typographically correct output, converting straight
quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to
ellipses.}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
it's size).}

\item{fill_page}{\code{TRUE} to layout the dashboard so that it fills the
browser to it's edges. \code{FALSE} to layout the dashboard as a
conventional scrolling web page.}

\item{orientation}{Determines whether level 2 headings are treated as
dashboard rows or dashboard columns.}

\item{theme}{Visual theme ("default", "bootstrap", "cerulean", "journal",
"flatly", "readable", "spacelab", "united", "cosmo", "lumen", "paper",
"sandstone", "simplex", or "yeti"). The "cosmo" theme is used when "default"
is specified.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
and "haddock". Pass NULL to prevent syntax highlighting.}

\item{mathjax}{Include mathjax. The "default" option uses an https URL from
the official MathJax CDN. The "local" option uses a local version of MathJax
(which is copied into the output directory). You can pass an alternate URL
or pass \code{NULL} to exclude MathJax entirely.}

\item{css}{One or more css files to include}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{lib_dir}{Directory to copy dependent HTML libraries (e.g. jquery,
bootstrap, etc.) into. By default this will be the name of the document with
\code{_files} appended to it.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{devel}{Enable development mode (used for development of the format
itself, not useful for users of the format).}

\item{...}{Unused}
}
\description{
Format for converting an R Markdown document to a grid oriented dashboard
layout. The dashboard flexibly adapts the size of it's plots and htmlwidgets
to its containing web page.
}
\details{
See the flexdashboard website for additional details on using
 flex_dashboard: http://rstudio.github.io/flexdashboard
}

