% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{flexdashboard-package}
\alias{flexdashboard}
\alias{flexdashboard-package}
\title{flexdashboard: Interactive dashboards for R}
\description{
Create interactive dashboards using \pkg{rmarkdown}.
}
\details{
\itemize{
\item{Use R Markdown to publish a group of related data visualizations as a dashboard.}
\item{Ideal for publishing interactive JavaScript visualizations based on htmlwidgets (also works with standard base, lattice, and grid graphics).}
\item{Flexible and easy to specify layouts. Charts are intelligently re-sized to fill the browser and adapted for display on mobile devices.}
\item{Optionally use Shiny to drive visualizations dynamically.}
}

See the flexdashboard website for additional documentation:
  \href{https://pkgs.rstudio.com/flexdashboard/}{https://pkgs.rstudio.com/flexdashboard/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkgs.rstudio.com/flexdashboard/}
  \item \url{https://github.com/rstudio/flexdashboard/}
  \item Report bugs at \url{https://github.com/rstudio/flexdashboard/issues}
}

}
\author{
\strong{Maintainer}: Carson Sievert \email{carson@rstudio.com} (\href{https://orcid.org/0000-0002-4958-2844}{ORCID})

Authors:
\itemize{
  \item Richard Iannone \email{rich@rstudio.com} (\href{https://orcid.org/0000-0003-3925-190X}{ORCID})
  \item JJ Allaire \email{jj@rstudio.com}
  \item Barbara Borges \email{barb.b.ribeiro@gmail.com}
}

Other contributors:
\itemize{
  \item  RStudio [copyright holder]
  \item  Keen IO (Dashboard CSS) [contributor, copyright holder]
  \item Abdullah Almsaeed (Dashboard CSS) [contributor, copyright holder]
  \item Jonas Mosbech (StickyTableHeaders) [contributor, copyright holder]
  \item Noel Bossart (Featherlight) [contributor, copyright holder]
  \item Lea Verou (Prism) [contributor, copyright holder]
  \item Dmitry Baranovskiy (Raphael.js) [contributor, copyright holder]
  \item  Sencha Labs (Raphael.js) [contributor, copyright holder]
  \item Bojan Djuricic (JustGage) [contributor, copyright holder]
  \item Tomas Sardyha (Sly) [contributor, copyright holder]
  \item Bryan Lewis (Examples) [contributor, copyright holder]
  \item Joshua Kunst (Examples) [contributor, copyright holder]
  \item Ryan Hafen (Examples) [contributor, copyright holder]
  \item Bob Rudis (Examples) [contributor, copyright holder]
  \item Joe Cheng (Examples) [contributor]
}

}
