% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{flexdashboard}
\alias{flexdashboard}
\title{flexdashboard: Interactive dashboards for R}
\description{
Create interactive dashboards using \pkg{rmarkdown}.
}
\details{
\itemize{
\item{Use R Markdown to publish a group of related data visualizations as a dashboard.}
\item{Ideal for publishing interactive JavaScript visualizations based on htmlwidgets (also works with standard base, lattice, and grid graphics).}
\item{Flexible and easy to specify layouts. Charts are intelligently re-sized to fill the browser and adapted for display on mobile devices.}
\item{Optionally use Shiny to drive visualizations dynamically.}
}

See the flexdashboard website for additional documentation:
\href{http://rmarkdown.rstudio.com/flexdashboard/}{http://rmarkdown.rstudio.com/flexdashboard/}
}
