% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuebox.R
\name{valueBox}
\alias{valueBox}
\title{Create a value box component for a dashboard.}
\usage{
valueBox(value, caption = NULL, icon = NULL, color = NULL)
}
\arguments{
\item{value}{The value to display in the box. Usually a number or short text.}

\item{caption}{The caption to display beneath the value.}

\item{icon}{An icon for the box (e.g. "fa-comments")}

\item{color}{Background color for the box. This can be one of the built-in
background colors ("primary", "info", "success", "warning", "danger") or
any valid CSS color value.}
}
\description{
A value box displays a value (usually a number) in large text, with a smaller
caption beneath, and a large icon on the right side.
}

