% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_dashboard.R
\name{flex_dashboard}
\alias{flex_dashboard}
\title{R Markdown Format for Flexible Dashboards}
\usage{
flex_dashboard(fig_width = 6, fig_height = 5, dev = "png", smart = TRUE,
  self_contained = TRUE, social = NULL, source_code = NULL,
  navbar = NULL, orientation = c("columns", "rows"),
  vertical_layout = c("fill", "scroll"), theme = "default",
  highlight = "default", mathjax = "default", extra_dependencies = NULL,
  css = NULL, includes = NULL, lib_dir = NULL, md_extensions = NULL,
  pandoc_args = NULL, devel = FALSE, ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{smart}{Produce typographically correct output, converting straight
quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to
ellipses.}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
it's size).}

\item{social}{Specify a character vector of social sharing services to
automatically add sharing links for them on the \code{navbar}. Valid values
are "twitter", "facebook", "google-plus", "linkedin", and "pinterest" (more
than one service can be specified).}

\item{source_code}{URL for source code of dashboard (used primarily for
publishing flexdashboard examples). Automatically creates a \code{navbar}
item which links to the source code.}

\item{navbar}{Optional list of elements to be placed on the flexdashboard
navigation bar. Each element should be a list containing a \code{title}
and/or \code{icon} field, a \code{url} field and an optional \code{align}
("left" or "right") field.}

\item{orientation}{Determines whether level 2 headings are treated as
dashboard rows or dashboard columns.}

\item{vertical_layout}{Vertical layout behavior: "fill" to vertically resize
charts so they completely fill the page; "scroll" to layout charts at their
natural height, scrolling the page if necessary.}

\item{theme}{Visual theme ("default", "bootstrap", "cerulean", "journal",
"flatly", "readable", "spacelab", "united", "cosmo", "lumen", "paper",
"sandstone", "simplex", or "yeti"). The "cosmo" theme is used when "default"
is specified.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
and "haddock". Pass NULL to prevent syntax highlighting.}

\item{mathjax}{Include mathjax. The "default" option uses an https URL from
the official MathJax CDN. The "local" option uses a local version of MathJax
(which is copied into the output directory). You can pass an alternate URL
or pass \code{NULL} to exclude MathJax entirely.}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}

\item{css}{One or more css files to include}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{lib_dir}{Directory to copy dependent HTML libraries (e.g. jquery,
bootstrap, etc.) into. By default this will be the name of the document with
\code{_files} appended to it.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{devel}{Enable development mode (used for development of the format
itself, not useful for users of the format).}

\item{...}{Unused}
}
\description{
Format for converting an R Markdown document to a grid oriented dashboard
layout. The dashboard flexibly adapts the size of it's plots and htmlwidgets
to its containing web page.
}
\details{
See the flexdashboard website for additional documentation
 flex_dashboard: http://rstudio.github.io/flexdashboard
}
\examples{
\dontrun{

library(rmarkdown)
library(flexdashboard)

# simple invocation
render("dashboard.Rmd", flex_dashboard())

# specify the theme option
render("pres.Rmd", flex_dashboard(theme = "yeti"))
}


}

