% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuebox.R
\name{valueBox}
\alias{renderValueBox}
\alias{valueBox}
\title{Create a value box for the main body of a dashboard.}
\usage{
valueBox(value, caption = NULL, icon = NULL, color = NULL)

renderValueBox(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{value}{The value to display in the box. Usually a number or short text.}

\item{caption}{The caption to display beneath the value.}

\item{icon}{An icon for the box (e.g. "fa-comments")}

\item{color}{Background color for the box. This can be one of the built-in
background colors ("primary", "info", "success", "warning", "danger") or
any valid CSS color value.}

\item{expr}{An expression that generates a value box}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
A value box displays a value (usually a number) in large text, with a smaller
caption beneath, and a large icon on the right side.
}

